+++ test/io_tests/request_response_tests/session_test.py
@@ -31,17 +31,14 @@ class TestSession:
 
         # Expected KNX/IP-Frame:
         exp_knxipframe = KNXIPFrame.init_from_body(
-            SessionRequest(
-                xknx,
-                ecdh_client_public_key=ecdh_public_key,
-            )
+            SessionRequest(ecdh_client_public_key=ecdh_public_key)
         )
 
         await session.start()
         transport_mock.send.assert_called_with(exp_knxipframe)
 
         # Response KNX/IP-Frame with wrong type
-        wrong_knxipframe = KNXIPFrame(xknx)
+        wrong_knxipframe = KNXIPFrame()
         wrong_knxipframe.init(KNXIPServiceType.SESSION_STATUS)
         with patch("logging.Logger.warning") as mock_warning:
             session.response_rec_callback(wrong_knxipframe, HPAI(), None)
@@ -49,7 +46,7 @@ class TestSession:
             assert session.success is False
 
         # Correct Response KNX/IP-Frame:
-        res_knxipframe = KNXIPFrame(xknx)
+        res_knxipframe = KNXIPFrame()
         res_knxipframe.init(KNXIPServiceType.SESSION_RESPONSE)
         res_knxipframe.body.secure_session_id = 5
         session.response_rec_callback(res_knxipframe, HPAI(), None)