+++ test/io_tests/tunnel_test.py
@@ -91,12 +91,9 @@ class TestUDPTunnel:
         self.tunnel.communication_channel = 1
 
         test_telegram = Telegram(payload=GroupValueWrite(DPTArray((1,))))
-        test_ack = KNXIPFrame.init_from_body(
-            TunnellingAck(self.xknx, sequence_counter=23)
-        )
+        test_ack = KNXIPFrame.init_from_body(TunnellingAck(sequence_counter=23))
         confirmation = KNXIPFrame.init_from_body(
             TunnellingRequest(
-                self.xknx,
                 communication_channel_id=1,
                 sequence_counter=23,
                 cemi=CEMIFrame.init_from_telegram(
@@ -141,7 +138,6 @@ class TestUDPTunnel:
 
         # Connect
         connect_request = ConnectRequest(
-            self.xknx,
             control_endpoint=local_endpoint,
             data_endpoint=local_endpoint,
         )
@@ -154,7 +150,6 @@ class TestUDPTunnel:
 
         connect_response_frame = KNXIPFrame.init_from_body(
             ConnectResponse(
-                self.xknx,
                 communication_channel=23,
                 data_endpoint=gateway_data_endpoint,
                 identifier=7,
@@ -170,7 +165,6 @@ class TestUDPTunnel:
         test_telegram = Telegram(payload=GroupValueWrite(DPTArray((1,))))
         test_telegram_frame = KNXIPFrame.init_from_body(
             TunnellingRequest(
-                self.xknx,
                 communication_channel_id=23,
                 sequence_counter=0,
                 cemi=CEMIFrame.init_from_telegram(
@@ -190,7 +184,8 @@ class TestUDPTunnel:
         # Disconnect
         self.tunnel.transport.send.reset_mock()
         disconnect_request = DisconnectRequest(
-            self.xknx, communication_channel_id=23, control_endpoint=local_endpoint
+            communication_channel_id=23,
+            control_endpoint=local_endpoint,
         )
         disconnect_frame = KNXIPFrame.init_from_body(disconnect_request)
 
@@ -199,10 +194,7 @@ class TestUDPTunnel:
         self.tunnel.transport.send.assert_called_once_with(disconnect_frame)
 
         disconnect_response_frame = KNXIPFrame.init_from_body(
-            DisconnectResponse(
-                self.xknx,
-                communication_channel_id=23,
-            )
+            DisconnectResponse(communication_channel_id=23)
         )
         self.tunnel.transport.handle_knxipframe(disconnect_response_frame, remote_addr)
         await disconnection_task
@@ -216,9 +208,9 @@ class TestUDPTunnel:
         self.tunnel.transport.getsockname = Mock(return_value=local_addr)
 
         description_request = KNXIPFrame.init_from_body(
-            DescriptionRequest(self.xknx, control_endpoint=self.tunnel.local_hpai)
+            DescriptionRequest(control_endpoint=self.tunnel.local_hpai)
         )
-        description_response = KNXIPFrame.init_from_body(DescriptionResponse(self.xknx))
+        description_response = KNXIPFrame.init_from_body(DescriptionResponse())
 
         task = asyncio.create_task(self.tunnel.request_description())
         await time_travel(0)