+++ test/knxip_tests/connect_request_test.py
@@ -1,7 +1,6 @@
 """Unit test for KNX/IP ConnectRequests."""
 import pytest
 
-from xknx import XKNX
 from xknx.exceptions import CouldNotParseKNXIP
 from xknx.knxip import HPAI, ConnectRequest, ConnectRequestType, KNXIPFrame
 
@@ -15,8 +14,7 @@ class TestKNXIPConnectRequest:
             "06 10 02 05 00 1A 08 01 C0 A8 2A 01 84 95 08 01"
             "C0 A8 2A 01 CC A9 04 04 02 00"
         )
-        xknx = XKNX()
-        knxipframe = KNXIPFrame(xknx)
+        knxipframe = KNXIPFrame()
         knxipframe.from_knx(raw)
 
         assert isinstance(knxipframe.body, ConnectRequest)
@@ -27,7 +25,6 @@ class TestKNXIPConnectRequest:
         assert knxipframe.body.data_endpoint == HPAI(ip_addr="192.168.42.1", port=52393)
 
         connect_request = ConnectRequest(
-            xknx,
             request_type=ConnectRequestType.TUNNEL_CONNECTION,
             control_endpoint=HPAI(ip_addr="192.168.42.1", port=33941),
             data_endpoint=HPAI(ip_addr="192.168.42.1", port=52393),
@@ -42,8 +39,7 @@ class TestKNXIPConnectRequest:
             "06 10 02 05 00 1A 08 01 C0 A8 2A 01 84 95 08 01"
             "C0 A8 2A 01 CC A9 02 04 02 00"
         )
-        xknx = XKNX()
-        knxipframe = KNXIPFrame(xknx)
+        knxipframe = KNXIPFrame()
         with pytest.raises(CouldNotParseKNXIP):
             knxipframe.from_knx(raw)
 
@@ -53,7 +49,6 @@ class TestKNXIPConnectRequest:
             "06 10 02 05 00 1A 08 01 C0 A8 2A 01 84 95 08 01"
             "C0 A8 2A 01 CC A9 04 04 02"
         )
-        xknx = XKNX()
-        knxipframe = KNXIPFrame(xknx)
+        knxipframe = KNXIPFrame()
         with pytest.raises(CouldNotParseKNXIP):
             knxipframe.from_knx(raw)