+++ test/knxip_tests/description_request_test.py
@@ -1,5 +1,4 @@
 """Unit test for KNX/IP DescriptionRequest objects."""
-from xknx import XKNX
 from xknx.knxip import HPAI, DescriptionRequest, KNXIPFrame
 
 
@@ -9,17 +8,14 @@ class TestKNXIPDescriptionRequest:
     def test_description_request(self):
         """Test parsing and streaming DescriptionRequest KNX/IP packet."""
         raw = bytes.fromhex("06 10 02 03 00 0E 08 01 7F 00 00 02 0E 57")
-        xknx = XKNX()
-        knxipframe = KNXIPFrame(xknx)
+        knxipframe = KNXIPFrame()
         knxipframe.from_knx(raw)
 
         assert isinstance(knxipframe.body, DescriptionRequest)
         assert knxipframe.body.control_endpoint == HPAI(ip_addr="127.0.0.2", port=3671)
 
         knxipframe2 = KNXIPFrame.init_from_body(
-            DescriptionRequest(
-                xknx, control_endpoint=HPAI(ip_addr="127.0.0.2", port=3671)
-            )
+            DescriptionRequest(control_endpoint=HPAI(ip_addr="127.0.0.2", port=3671))
         )
 
         assert knxipframe2.to_knx() == raw