+++ test/knxip_tests/disconnect_request_test.py
@@ -1,7 +1,6 @@
 """Unit test for KNX/IP DisconnectRequest objects."""
 import pytest
 
-from xknx import XKNX
 from xknx.exceptions import CouldNotParseKNXIP
 from xknx.knxip import HPAI, DisconnectRequest, KNXIPFrame
 
@@ -12,8 +11,7 @@ class TestKNXIPDisconnectRequest:
     def test_disconnect_request(self):
         """Test parsing and streaming DisconnectRequest KNX/IP packet."""
         raw = bytes.fromhex("06 10 02 09 00 10 15 00 08 01 C0 A8 C8 0C C3 B4")
-        xknx = XKNX()
-        knxipframe = KNXIPFrame(xknx)
+        knxipframe = KNXIPFrame()
         knxipframe.from_knx(raw)
 
         assert isinstance(knxipframe.body, DisconnectRequest)
@@ -24,7 +22,6 @@ class TestKNXIPDisconnectRequest:
         )
 
         disconnect_request = DisconnectRequest(
-            xknx,
             communication_channel_id=21,
             control_endpoint=HPAI(ip_addr="192.168.200.12", port=50100),
         )
@@ -35,7 +32,6 @@ class TestKNXIPDisconnectRequest:
     def test_from_knx_wrong_length(self):
         """Test parsing and streaming wrong DisconnectRequest."""
         raw = bytes((0x06, 0x10, 0x02, 0x09, 0x00, 0x10))
-        xknx = XKNX()
-        knxipframe = KNXIPFrame(xknx)
+        knxipframe = KNXIPFrame()
         with pytest.raises(CouldNotParseKNXIP):
             knxipframe.from_knx(raw)