+++ test/knxip_tests/search_request_test.py
@@ -1,5 +1,4 @@
 """Unit test for KNX/IP SearchRequest objects."""
-from xknx import XKNX
 from xknx.knxip import HPAI, KNXIPFrame, SearchRequest
 
 
@@ -9,8 +8,7 @@ class TestKNXIPSearchRequest:
     def test_search_request(self):
         """Test parsing and streaming SearchRequest KNX/IP packet."""
         raw = bytes.fromhex("06 10 02 01 00 0E 08 01 E0 00 17 0C 0E 57")
-        xknx = XKNX()
-        knxipframe = KNXIPFrame(xknx)
+        knxipframe = KNXIPFrame()
         knxipframe.from_knx(raw)
 
         assert isinstance(knxipframe.body, SearchRequest)
@@ -19,9 +17,7 @@ class TestKNXIPSearchRequest:
         )
 
         knxipframe2 = KNXIPFrame.init_from_body(
-            SearchRequest(
-                xknx, discovery_endpoint=HPAI(ip_addr="224.0.23.12", port=3671)
-            )
+            SearchRequest(discovery_endpoint=HPAI(ip_addr="224.0.23.12", port=3671))
         )
 
         assert knxipframe2.to_knx() == raw