+++ changelog.md
@@ -5,7 +5,7 @@
 ### Internal
 
 - Move `levels` instance attribute form `GroupAddress` to `address_format` class variable
-- Remove xknx form CEMIFrame class
+- Remove xknx form every class in the knxip package: CEMIFrame, KNXIPFrame and KNXIPBody (and subclasses)
 
 ## 0.20.3 Threading fixes 2022-04-15